reformat to accept yaml file as input

+++ processing/manual_correction.sh
@@ -2,29 +2,70 @@
 
 # Run this script in results/data folder
 #
-# USAGE - define $FILES variable as list of segmentations (separeted by space)
-# to correct in the terminal and then run:
-#   PATH_TO_SPINEGENERIC/processing/manual_correction.sh $FILES
+# USAGE:
+# Define files.yaml file as following:
+#   ---
+#   FILES_SEG:
+#   - sub-amu01_T1w_RPI_r.nii.gz
+#   - sub-amu01_T2w_RPI_r.nii.gz
+#   - sub-brnoCeitec01_T2star_rms.nii.gz
+#   FILES_LABEL:
+#   - sub-amu01
+#   - sub-brnoUhb01
 #
+#
+# THEN RUN:
+#   PATH_TO_SPINEGENERIC/processing/manual_correction.sh files.yaml
+#
+# Authors: Julien Cohen-Adad, Jan Valosek
 
 
-# Print help
+# function for parsing input yaml file
+yaml_parser() {
+
+  INPUT_YAML=$1
+
+  # tr: delete all newlines begining with dash (-), so convert input file into one line without dashes
+  # (unfortunately this also delete all other dashes, so it is neccesary to use 3rd sed)
+  # 1st sed: delete "FILES_LABEL:" string and everything after it
+  # 2nd sed: delete "FILES_SEG:" string itself
+  # 3rd sed: replace "sub" by "sub-" (tr command deleted all dashes (-))
+  FILES_SEG=$(cat $INPUT_YAML | tr -d '\n-' | sed 's/FILES_LABEL:.*//' | sed 's/FILES_SEG://' | sed 's/sub/sub-/g')
+  # tr: delete all newlines begining with dash (-), so convert input file into one line without dashes
+  # (unfortunately this also delete all other dashes, so it is neccesary to use 3rd sed)
+  # 1st sed: delete everything before "FILES_LABEL:" string including string "FILES_LABEL:" itself
+  # 2nd sed: replace "sub" by "sub-" (tr command deleted all dashes (-))
+  FILES_LABEL=$(cat $INPUT_YAML | tr -d '\n-' | sed 's/.*FILES_LABEL://' | sed 's/sub/sub-/g')
+}
+
+# Print help, if invalid input
 if [[ $# != 1 ]] || [[ $1 == "--help" ]] || [[ $1 == "-h" ]];then
 
-  echo "Invalid input. Please define \$FILES variable as list of segmentations (separeted by space) to correct in the terminal."
-  echo -e "Example:\n\tFILES=(sub-amu02_acq-T1w_MTS.nii.gz sub-beijingGE04_T2w_RPI_r.nii.gz)"
-  echo -e "Then run:\n\tPATH_TO_SPINEGENERIC/processing/manual_correction.sh \$FILES"
+  echo -e "Invalid input. \n\nPlease define files.yaml file with list of segmentations to correct"
+  echo -e "and list of subjects to correct labeling on (one file perline)."
+  echo -e "\nExample:"
+  echo -e "\t---"
+  echo -e "\tFILES_SEG:"
+  echo -e "\t- sub-amu01_T1w_RPI_r.nii.gz"
+  echo -e "\t- sub-amu01_T2w_RPI_r.nii.gz"
+  echo -e "\t- sub-brnoCeitec01_T2star_rms.nii.gz"
+  echo -e "\tFILES_LABEL:"
+  echo -e "\t- sub-amu01"
+  echo -e "\t- sub-brnoUhb01"
+  echo -e "\nThen run this script in results/data folder:\n\tPATH_TO_SPINEGENERIC/processing/manual_correction.sh files.yaml"
 
 else
 
-  # Folder to output the manual labels
+  # call yaml_parser function
+  yaml_parser $1
+
+  # Folder to output the manual segmentations and labels
   PATH_SEGMANUAL="../../seg_manual"
   mkdir $PATH_SEGMANUAL
-  # List of files to correct segmentation on
-  FILES=$1
 
-  # Loop across files
-  for file in ${FILES[@]}; do
+  # Loop across segmentation files
+  for file in $FILES_SEG; do
+
     # extract subject using first delimiter '_'
     subject=${file%%_*}
     # check if file is under dwi/ or anat/ folder and get fname_data
@@ -41,11 +82,33 @@ else
       fname_seg=${fname_data%%".nii.gz"*}_seg.nii.gz${fname_data##*".nii.gz"}
       fname_seg_dest=${PATH_SEGMANUAL}/${file%%".nii.gz"*}_seg-manual.nii.gz${file##*".nii.gz"}
     fi
-    # Copy file to PATH_SEGMANUAL
-    cp $fname_seg $fname_seg_dest
-    # Launch FSLeyes
-    echo "In FSLeyes, click on 'Edit mode', correct the segmentation, then save it with the same name (overwrite)."
-    fsleyes -yh $fname_data $fname_seg_dest -cm red
+    # check if segmentation file exist, i.e., filename is correct
+    if [[ -f $fname_seg ]];then
+      # Copy file to PATH_SEGMANUAL
+      cp $fname_seg $fname_seg_dest
+      # Launch FSLeyes
+      echo "In FSLeyes, click on 'Edit mode', correct the segmentation, then save it with the same name (overwrite)."
+      fsleyes -yh $fname_data $fname_seg_dest -cm red
+    else
+      echo "File $file does not exist. Please verity if you entered filename correctly."
+    fi
+
+  done
+
+  # Loop across labeling files
+  for subject in $FILES_LABEL; do
+
+    fname_label=$subject/anat/${subject}_T1w_RPI_r.nii.gz
+    fname_label_dest=${PATH_SEGMANUAL}/${subject}_T1w_RPI_r_labels-manual.nii.gz
+    # check if labeling file exist, i.e., filename is correct
+    if [[ -f $fname_label ]];then
+      # Launch GUI for manual labeling
+      echo "In sct_label_utils GUI, select C3 and C5, then click 'Save and Quit'."
+      sct_label_utils -i $fname_label -create-viewer 3,5 -o $fname_label_dest
+    else
+      echo "File $fname_label does not exist. Please verity if you entered filename correctly."
+    fi
+
   done
 
 fi
